backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
pydantic
aiofiles
python-multipart
openai-whisper
torch
transformers
sentence-transformers
requests
websockets
